+++ lib/trackdown/location_result.rb
@@ -4,13 +4,26 @@ require 'countries'
 
 module Trackdown
   class LocationResult
-    attr_reader :country_code, :country_name, :city, :flag_emoji
+    attr_reader :country_code, :country_name, :city, :flag_emoji,
+                :region, :region_code, :continent, :timezone, :latitude, :longitude,
+                :postal_code, :metro_code
 
-    def initialize(country_code, country_name, city, flag_emoji)
+    def initialize(country_code, country_name, city, flag_emoji,
+                   region: nil, region_code: nil, continent: nil,
+                   timezone: nil, latitude: nil, longitude: nil,
+                   postal_code: nil, metro_code: nil)
       @country_code = country_code
       @country_name = country_name
       @city = city
       @flag_emoji = flag_emoji
+      @region = region
+      @region_code = region_code
+      @continent = continent
+      @timezone = timezone
+      @latitude = latitude
+      @longitude = longitude
+      @postal_code = postal_code
+      @metro_code = metro_code
     end
 
     alias_method :country, :country_name
@@ -29,6 +42,14 @@ module Trackdown
         country_name: @country_name,
         city: @city,
         flag_emoji: @flag_emoji,
+        region: @region,
+        region_code: @region_code,
+        continent: @continent,
+        timezone: @timezone,
+        latitude: @latitude,
+        longitude: @longitude,
+        postal_code: @postal_code,
+        metro_code: @metro_code,
         country_info: country_info&.data || {}
       }
     end